+++ src/main/java/chess/domain/chessPiece/Bishop.java
@@ -1,8 +1,7 @@
 package chess.domain.chessPiece;
 
-import chess.domain.Direction;
-import chess.domain.Position;
-import chess.domain.Team;
+import chess.domain.position.Direction;
+import chess.domain.position.Position;
 
 import java.util.ArrayList;
 import java.util.Collections;
