+++ src/main/java/chess/domain/chessPiece/King.java
@@ -1,6 +1,7 @@
 package chess.domain.chessPiece;
 
-import chess.domain.*;
+import chess.domain.position.Direction;
+import chess.domain.position.Position;
 
 import java.util.ArrayList;
 import java.util.Collections;
