+++ src/main/java/chess/domain/chessPiece/Piece.java
@@ -1,7 +1,6 @@
 package chess.domain.chessPiece;
 
-import chess.domain.Position;
-import chess.domain.Team;
+import chess.domain.position.Position;
 
 import java.util.List;
