+++ src/test/java/chess/domain/DirectionTest.java
@@ -1,7 +1,7 @@
 package chess.domain;
 
-import chess.domain.Direction;
-import chess.domain.Position;
+import chess.domain.position.Direction;
+import chess.domain.position.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
