+++ src/test/java/chess/domain/RowTest.java
@@ -1,6 +1,6 @@
 package chess.domain;
 
-import chess.domain.Row;
+import chess.domain.position.Row;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
