+++ src/test/java/chess/domain/chessPiece/QueenTest.java
@@ -1,16 +1,15 @@
 package chess.domain.chessPiece;
 
-import chess.domain.Position;
+import chess.domain.position.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
-import static chess.domain.Team.BLACK;
-import static chess.domain.Team.WHITE;
+import static chess.domain.chessPiece.Team.BLACK;
+import static chess.domain.chessPiece.Team.WHITE;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
 
 class QueenTest {
     @Test
